+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.2.0
 
 * Change initialization method, add new getBlackBoxAsync method
+* Add showCaptcha function
 ## 1.1.0
 
 * Optimize and modify configuration